[BUGFIX] Use correct path to JS resources
This problem is related to the change https://review.typo3.org/48915/.
The patch removed usage of ExtensionManagementUtility::extRelPath and
replaced it with other calls. While trying to resolve a relative path
in BackendController::loadResourcesForRegisteredNavigationComponents,
a wrong replacement was introduced. This patch ensures the old behaviour
to retrieve the relative path.

Resolves: #78490
Releases: master
Change-Id: I2c2edf6d1b02efd95b4485da88777eec0dd515c6
Reviewed-on: https://review.typo3.org/50475
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Malte Koitka <malte@cytrus.de>
Tested-by: Malte Koitka <malte@cytrus.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>