[BUGFIX] Opening be_groups records may throw exception
The new flex form data structure handling in more picky and throws
exception if a given data structure identifier can not be resolved.
The FormEngine logic that displays exclude fields of flex forms
has to do some guess work to find possible data structures. This
may fail, resulting in an exception.
The patch changes the according general exception to a specific one,
catches it in the find flex form access rights logic, and adds
more docu on what exactly is happening in this code.

Change-Id: I2b808cbe804ba4ab967c8d95b17c0a12ec65ffd4
Resolves: #78852
Related: #78581
Releases: master
Reviewed-on: https://review.typo3.org/50820
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Alexander Stehlik <alexander.stehlik@gmail.com>
Tested-by: Alexander Stehlik <alexander.stehlik@gmail.com>
Reviewed-by: Alexander Rothmund <alexander@rothmund.ch>
Tested-by: Alexander Rothmund <alexander@rothmund.ch>
Reviewed-by: Jan Helke <typo3@helke.de>
Tested-by: Jan Helke <typo3@helke.de>