[BUGFIX] FormEngine group element top/bottom buttons
The sort top/bottom buttons in group element no longer depend
on the number of selected items. This is inconvenient anyway
since these sorting buttons only appeared after saving when
a bunch of items was selected, but not immediately.
The patch additionally fixes a failing acceptance test.

Change-Id: Ice0d179e7c74c3a1227c727c672b40844d6852c8
Resolves: #79149
Related: #78899
Releases: master
Reviewed-on: https://review.typo3.org/51111
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Jan Helke <typo3@helke.de>
Tested-by: Jan Helke <typo3@helke.de>