[BUGFIX] Show record title for inline element
Currently no record titles for nested inline elements are shown. Due to
the current method to resolve a record title for an inline element
always an empty title is returned as there is the same processing as for
select fields. This patch introduces an own resolving method according to
the already resolved children attached to the current field.

Resolves: #78726
Releases: master, 7.6
Change-Id: I7f62ad843552534b36c0ccd91488c79ff796ec0c
Reviewed-on: https://review.typo3.org/51115
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Josef Glatz <josef.glatz@typo3.org>
Tested-by: Josef Glatz <josef.glatz@typo3.org>
Reviewed-by: Helmut Hummel <typo3@helhum.io>
Tested-by: Helmut Hummel <typo3@helhum.io>