[TASK] Replace ExtDirect State for module menu with jQuery
The obsolete State Provider for storing and retrieving the information 
if the backend module menu should be collapsed or not
should not use ExtDirect but instead us the native Storage functionality
already in place for TYPO3.

The namespace (BackendComponents.States.typo3-module-menu) is
kept for now (see BackendController.php for details on ExtDirect), but could
easily be modified.

Resolves: #79211
Releases: master
Change-Id: I4de6b46ea0f18201cb8bb0af6077c1d9dfffcec7
Reviewed-on: https://review.typo3.org/51208
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>