[TASK] Use “error” icon for validation failures in FormEngine
Currently, a warning sign is rendered next to field labels whose
validation failed. Failed validations are errors and thus another icon
should be rendered. With this patch, a red circle containing an
exclamation mark is rendered. It’s the same icon as in the tabs holding
affected fields.

Resolves: #79537
Releases: master
Change-Id: I01b783c36dfac551bf6e9f0cb8fc703290c3eb72
Reviewed-on: https://review.typo3.org/51460
Tested-by: TYPO3com <no-reply@typo3.com>
Tested-by: Riccardo De Contardi <erredeco@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Thomas Hohn <thomas@hohn.dk>
Tested-by: Thomas Hohn <thomas@hohn.dk>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>