[BUGFIX] PageTree: Re-introduce queueing
With the removal of the ExtJS state providers, the polyfill
`TYPO3ExtJSStateProviderBridge` with a limited functional subset was
introduced. Previously, changes were queued to reduce the amount of
requests, which got lost since then. This was a complex workaround in
first place and the real issue that the same data is sent *per node*
lies somewhere deep in ExtJS and was never fixed.

As we aim to get rid of the last pieces of ExtJS sooner than later, the
queueing mechanism is added again keep the load low and to worship our
servers.

Resolves: #79693
Related: #79227
Releases: master
Change-Id: Ibcccd9be183437192027cbfab634515f710728c6
Reviewed-on: https://review.typo3.org/51591
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Thomas Hohn <thomas@hohn.dk>
Tested-by: Thomas Hohn <thomas@hohn.dk>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Mona Muzaffar <mona.muzaffar@gmx.de>
Reviewed-by: Markus Sommer <markussom@posteo.de>
Tested-by: Markus Sommer <markussom@posteo.de>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>