[BUGFIX] Flex form section container with group elements
The "create new flex form section container" ajax request fails to set
initialized data values of the data compiler to the render engine.
As a result, data values that expect to be at least an empty array
are null and for instance the GroupElement stumbles. Properly assign
calculated container data to the render engine to fix this.

Change-Id: I1205c2814608bbb23d4842333f0b047bc356b02a
Resolves: #80222
Releases: master
Reviewed-on: https://review.typo3.org/52105
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Jan Helke <typo3@helke.de>
Tested-by: Jan Helke <typo3@helke.de>
Reviewed-by: Frank Nägler <frank.naegler@typo3.org>
Tested-by: Frank Nägler <frank.naegler@typo3.org>