[BUGFIX] JS: Invert dependency definition for FormEngine and Validation
FormEngineValidation is meant as integral part of FormEngine.
The current dependency definition is inverted, such that FormEngine
now depends on FormEngineValidation and not the other way around.
Note that FormEngineValidation does not need any other feature
of FormEngine, so this is safe.

This solves a possible race condition with other modules.

Releases: master, 7.6
Resolves: #80459
Change-Id: Ie2cb9dd90220b37a98bc8e3ca174fd074488cc8d
Reviewed-on: https://review.typo3.org/52178
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Reviewed-by: Frank Nägler <frank.naegler@typo3.org>
Tested-by: Frank Nägler <frank.naegler@typo3.org>
Reviewed-by: Kay Strobach <typo3@kay-strobach.de>
Tested-by: Kay Strobach <typo3@kay-strobach.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>