[BUGFIX] Avoid unnecessarily fetching file for processing
Files that are not configured as images, won't be processed,
but are still unnecessarily fetched for processing.

To avoid this, we change the processing API, that fetching
only happens, if the file is then used by following code.

In addition we avoid showing a non descriptive thumb
in FileInfoElement, which is used when editing meta data of the file.

Resolves: #81279
Releases: master, 8.7, 7.6
Change-Id: Ie68aa02b1b6149ddb7fbda346ac312d3646e260c
Reviewed-on: https://review.typo3.org/52874
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Philipp Gampe <philipp.gampe@typo3.org>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>
Reviewed-by: Christer V <cvi@systime.dk>
Reviewed-by: Frans Saris <franssaris@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>