[BUGFIX] Use localize command when translating page record
Currently, when using the dropdown to create a page translation within
the Page Module, DataHandler is called with just a simple "new" record
of type "pages_language_overlay".

However, a classic "localize" DataHandler command is more appropriate,
allowing to utilize several localization functionalities built in.

In order to achieve this, the option tags in the selector contain the full
URL to the DataHandler call, as well as the redirect to FormEngine (which
in turn redirects back to the current page module URL).

Resolves: #81345
Releases: master, 8.7
Change-Id: I740c2d1d11494963283bc14b1f534bd344d0764a
Reviewed-on: https://review.typo3.org/52945
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: TYPO3com <no-reply@typo3.com>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>