[BUGFIX] Module menu clicks with slow connections opens standalone module
If loading the backend using a slow internet connection an early click on
any item in the module menu on the left side causes the link to be opened
directly. The reason is, that JavaScript and thus click handlers are not
completely loaded yet. Thus, the regular DOM behavior of anchor tags is
executed.

To circumvent this behavior, href-arguments have to be disabled and the
module URL moved to a dedicated new data-attribute. As a result all items
that trigger a module change have to be adjusted as well.

Resolves: #80565
Releases: master, 8.7
Change-Id: I4687e6594eb539acc0d299dc8fc02e6820019911
Reviewed-on: https://review.typo3.org/52996
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>