[BUGFIX] Set anchor in all preview links of content elements
The button »View webpage« in the list module opens a popup
with a link to the current page and an anchor to the
selected content element. This behaviour is expected as well
while selecting the »Show« option in the context menu of
content elements or the »Save and view page« button in
edit forms. Both places didn't added the anchor.

Resolves: #81250
Releases: master, 8.7
Change-Id: Ifa8da940eb12820f9b0c54957928fd8d9f3703b5
Reviewed-on: https://review.typo3.org/53691
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>