[BUGFIX] Properly escape "dropzone-target" selector in DragUploader.js
We properly escape some characters of "dropzone-target" data attribute,
since it is being used as a CSS selector to insert the dropzone in
our DOM. The "dropzone-target" might contain characters that have a
special meaning in CSS, like for example a dot. Especially the dot
is typical for cases like flexforms fields.

This allows drag and drop file uploads even on such cases (like for
example working drag and drop file uploads in DCE content elements).

Resolves: #81812
Releases: master, 8.7
Change-Id: Ib1f5b5063e390f08436fd3a51978842754b698ef
Reviewed-on: https://review.typo3.org/53416
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>