[BUGFIX] Changed logical operator.
As the Exception says: It is not allowed to have a section without a type
OR a type without a section. So it is not allowed to have one without the
other, but it's allowed to have both or none of them. --> This is a
EXCLUSIVE OR situation.

Releases: master
Resolves: #81985
Change-Id: I00594b75c99b02d1950be106dc4c3389709af231
Reviewed-on: https://review.typo3.org/53610
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Jonathan IROULIN <joniroutypo3@gmail.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>