[FEATURE] Add hook for used/unused content record processing
This patch adds a new hook to determine whether a content record is
used or not. Since the default decision is fairly basic (only checks if
$columns contains $record['colPos']) a hook makes it possible for third
party extensions to change the decision and validate on their own.

Included rst file shows how to register to the hook and provides example
code how to determine the result.

Resolves: #82213
Releases: master
Change-Id: Id338254dae6575ba7b91141c10a42b99401614c9
Reviewed-on: https://review.typo3.org/53811
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Claus Due <claus@phpmind.net>
Reviewed-by: Henning Liebe <h.liebe@neusta.de>
Tested-by: Henning Liebe <h.liebe@neusta.de>
Reviewed-by: Joerg Boesche <typo3@joergboesche.de>
Tested-by: Nicole Cordes <typo3@cordes.co>
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Tested-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>