[TASK] Improve performance of UriBuilder by memory cache
The TYPO3 backend is eager to generate the same URL
with the same parameters multiple times, causing a lot
of cascading calls to URI parsing, token generation etc.

Turning the UriBuilder into a Singleton removes many
hundreds of calls to instantiate the object - and remembering
the URLs that were generated further saves many hundred
calls to UriBuilder->buildUri in for example the page module.

Change-Id: I7444ccf8bb27789c489f7d7c15c65449fa3456a0
Resolves: #82237
Releases: master
Reviewed-on: https://review.typo3.org/53827
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>