[TASK] Reduce PHP queries of pagetree SQL
The TYPO3 Backend PageTree, built in TYPO3 4.5, has
some strange quirks resolving DB records, doing
a lot of queries for fetching a single page multiple
times. This is unnecessary because this can be fetched
with one query, which happens anyway. Additionally,
the WSOL should only happen when a workspace is selected.

Explicitly querying for "-1" pid etc. should be handled
via SQL, and not explicitly implemented by the page tree.

Bottom line: Remove one SQL-query per page which is loaded.

As most of the PHP classes will be restructured with the upcoming
ExtJS / ExtDirect removal, the removed methods will be part
of the breaking change of ExtJS removal.

Resolves: #82945
Releases: master
Change-Id: I7b1d79b40d0e9212cc0884c9440e5725e4f74d8e
Reviewed-on: https://review.typo3.org/54574
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Tested-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>