[BUGIFX] Fix JS error in filtering page tree
While fixing #82877 a mistake was introduced which leads
to a JavaScript error when filtering the page tree.

Problem is that PageTree.select() is called with
the filter tree currently active, but this method
is only functional for the full page tree since quite
a while now.

We now avoid calling this method and document the behavior
properly in method annotations.

Additionally a typo in the variable name is fixed,
which lead to a jquery warning when selecting a page.

Resolves: #83039
Releases: master, 8.7
Change-Id: I6440fb7f6cfd6ead8b5b29dea7c5081dd688a5c3
Reviewed-on: https://review.typo3.org/54689
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>