[TASK] UX: Streamline Page Translation selector in Page Module
The page translation selector in the page module has the label outside
of the dropdown, where as the List module has the label as the first option.

This patch puts the label inside the first option of the dropdown selector
in the page module.

This way, the selector box is aligned to the left, and thus, does not look
like it's placed "somewhere" in the HTML code.

Resolves: #83271
Releases: master
Change-Id: Idbedf2f6cfb9056e0c89a69029f31b4b47673a02
Reviewed-on: https://review.typo3.org/55001
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>