[!!!][TASK] Remove EXT:backend/Resources/Private/Templates/Close.html
Remove unused file EXT:backend/Resources/Private/Templates/Close.html

In #68108 the file Close.html was erroneously moved to
EXT:backend/Resources/Private/Templates/ which is a path that is not
directly accessible for web users, but the file must be. In #83258 this
was fixed by copying the file to EXT:backend/Resources/Public/Html/ to
make this a non-breaking change back-portable down to 7.6.

Releases: master
Resolves: #83284
Related: #83258
Related: #68108
Change-Id: I4ece2b4d6069e0a573ecfa092c73b80d2c07a42f
Reviewed-on: https://review.typo3.org/55023
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Tested-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>