[TASK] Make stateless Bootstrap methods static
Methods that do not use $this but mainly act as utility
may be used statically.
The previous non-static chainable behaviour is preserved
by returning static::$instance. (Invoking a static method
non statically does not trigger E_NOTICE or alike.)

The motivation for this change is to deprecate the
Bootstrap instanciated usage at some (later) point
in order to reduce global state.

The following methods are affected:
 * initializeLanguageObject
 * initializeBackendAuthentication
 * initializeBackendUser
 * initializeBackendRouter
 * loadExtTables
 * loadBaseTca
 * loadTypo3LoadedExtAndExtLocalconf
 * unsetReservedGlobalVariables
 * startOutputBuffering
 * disableCoreCache

Releases: master
Resolves: #83952
Change-Id: I59d3027c5d10326d7bab2ae02e6ff0eb836f23e4
Reviewed-on: https://review.typo3.org/55775
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>