[BUGFIX] Unbind load event after page tree refresh
When an action in the page tree is called via the context menu that
enforces a reload of the page tree, the tree is reloaded every time once
the content container is refreshed.

Change the event binding to use `one` to unbind automatically once the
tree was reloaded.

Resolves: #84294
Related: #82970
Releases: 8.7
Change-Id: Ife69dfbceb602a0d610397f313ea539fd9b93711
Reviewed-on: https://review.typo3.org/56170
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Benjamin Kluge <b.kluge@neusta.de>
Tested-by: Benjamin Kluge <b.kluge@neusta.de>