[TASK] Use ServerRequestInterface in Wizard/EditController
This patch changes the EditController to consistently use
ServerRequestInterface instead of GeneralUtility and
HttpUtility.

The method main() is kept to keep this change non-breaking,
it translates the given response objects back to strings/redirects
which match existing behaviour.

Also, to prepare for refactoring in v10,
main() and the new protected properties will throw a
deprecation notice if called.

Resolves: #84327
Releases: master
Change-Id: I9bd055f4a81e5f9ca324954bc03206c9f463cb11
Reviewed-on: https://review.typo3.org/56232
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>